별점 등록 api

+++ src/backend/src/main/java/com/graduate/odondong/dto/BathroomRequestDto.java
@@ -12,5 +12,5 @@ public class BathroomRequestDto {
     private String address;
     private String addressDetail;
     private String imageUrl;
-    private Float rate;
+    private Double rate;
 }
